requirements.txt
Django~=5.1
openpyxl~=3.1.5
django-unfold
django-import-export
psycopg2-binary>=2.9.3
celery==5.4.0
redis==5.0.8
